refactor: modificar type de preco na tabela servico para BigDecimal

+++ src/main/java/com/example/barbearia/dto/request/ServicoRequestDto.java
@@ -11,6 +11,6 @@ public record ServicoRequestDto(
         @NotNull(message = "Duração é obrigatório")
         Integer duracao,
         @NotNull(message = "Preço é obrigatório")
-        Double preco
+        BigDecimal preco
 ) {
 }
